apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
redis[asyncio]==5.2.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.35.0
cryptography==43.0.3
apscheduler==3.10.4
opencv-python-headless==4.10.0.84
numpy==2.1.3
Pillow==11.0.0
psutil==6.1.0
PyYAML==6.0.2
shapely==2.0.6
